api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.7.4
pydantic-settings==2.3.4
boto3==1.34.131
python-multipart==0.0.9
httpx==0.27.0
biopython==1.84
openai==1.35.0
websockets==12.0
